feat(step2): wire chapter layouts into auto-authoring + shared layout module
Extracts the chapter spatial-layout logic into scripts/lib/chapter-layout.mjs
(layout / pickLayout / buildChapter), used by BOTH generators (DRY). The
auto-authoring tool now picks a layout by station count — pickLayout: 1 → linear,
2-3 → arc, 4+ → grid — so a chapter's camera tour matches its shape.

- gen-deck-layouts.mjs refactored onto the shared module (byte-identical output
  to the inline version it replaced — deck-lay-* unchanged, a clean regression
  signal).
- gen-deck-authored.mjs: chapters now use buildChapter + pickLayout; titles note
  the chosen layout. On the fill deck, Chapter 1 (2 slides) → arc, the rest → linear.

Verified real-browser GPU: deck-lay-arc (refactor) + deck-auth-ch0 (new arc
chapter) both render with 0 shader errors. Full suite 82/82.
